requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
jupyter>=1.0
python-dateutil>=2.8
shap>=0.44
fastapi>=0.104
uvicorn>=0.24
